mi_assessment/requirements.txt
Django==4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-crispy-forms
crispy-bootstrap5
